Imagefactory: TinMan
* Added ability to pass in overrides for oz configurations in the
format of a json/dict string. Example:

--parameter oz_overrides "{'libvirt': {'memory': 4096, 'image_type': 'kvm', 'cpus': 2}}"

Added python-simplejson Requires to imagefactory-plugins.spec.in

imgfac/FactoryUtils.py

* Added the check_qcow_size function into this common place.

imagefactory_plugins/vSphere/vSphere.py

* Now checks if "input" image is raw or qcow; if qcow2, it
converts to raw and passes the new raw image as the input
for the vSphere conversion

imagefactory_plugins/OpenStack/OpenStack.py

* Remove the check_qcow_size def, added import

imagefactory_plugins/ovfcommon/ovfcommon.py

* Remove the check_qcow_size def, added import
