+++ imagefactory_plugins/imagefactory-plugins.spec.in
@@ -78,6 +78,7 @@ License: ASL 2.0
 Requires: oz >= 0.12.0
 Requires: imagefactory-plugins
 Requires: imagefactory-plugin-api = 1.0
+Requires: python-simplejson
 
 %description TinMan
 An OS plugin to support Fedora OSes
