+++ imgfac/FactoryUtils.py
@@ -11,6 +11,7 @@ import re
 from imgfac.ImageFactoryException import ImageFactoryException
 import subprocess
 import logging
+import struct
 
 
 def launch_inspect_and_mount(diskfile, readonly=False):
@@ -233,3 +234,40 @@ def parameter_cast_to_bool(ival):
         if lower == 'yes' or lower == 'true' or lower == '1':
             return True
     return None
+
+def check_qcow_size(self, filename):
+    # Detect if an image is in qcow format
+    # If it is, return the size of the underlying disk image
+    # If it isn't, return None
+
+    # For interested parties, this is the QCOW header struct in C
+    # struct qcow_header {
+    #    uint32_t magic;
+    #    uint32_t version;
+    #    uint64_t backing_file_offset;
+    #    uint32_t backing_file_size;
+    #    uint32_t cluster_bits;
+    #    uint64_t size; /* in bytes */
+    #    uint32_t crypt_method;
+    #    uint32_t l1_size;
+    #    uint64_t l1_table_offset;
+    #    uint64_t refcount_table_offset;
+    #    uint32_t refcount_table_clusters;
+    #    uint32_t nb_snapshots;
+    #    uint64_t snapshots_offset;
+    # };
+
+    # And in Python struct format string-ese
+    qcow_struct=">IIQIIQIIQQIIQ" # > means big-endian
+    qcow_magic = 0x514649FB # 'Q' 'F' 'I' 0xFB
+
+    f = open(filename,"r")
+    pack = f.read(struct.calcsize(qcow_struct))
+    f.close()
+
+    unpack = struct.unpack(qcow_struct, pack)
+
+    if unpack[0] == qcow_magic:
+        return unpack[5]
+    else:
+        return None
